Lastnext/myLubd/src/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.8.30
charset-normalizer==3.4.0
defusedxml==0.7.1
Django==4.2.16
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
httpie==3.2.4
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
pillow==10.4.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
Pygments==2.18.0
PyJWT==2.9.0
PySocks==1.7.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
sqlparse==0.5.2
typing_extensions==4.12.2
urllib3==2.2.3
